Renamed SpringFeign to SpringClient
