+++ logger/src/main/kotlin/com/markklim/libs/ginger/dao/log/http/LogType.kt
@@ -11,8 +11,8 @@ enum class LogType {
     FEIGN_RESP,
     FEIGN_RESP_B,
 
-    SPRING_FEIGN_REQ,
-    SPRING_FEIGN_REQ_B,
-    SPRING_FEIGN_RESP,
-    SPRING_FEIGN_RESP_B
+    SPRING_CLIENT_REQ,
+    SPRING_CLIENT_REQ_B,
+    SPRING_CLIENT_RESP,
+    SPRING_CLIENT_RESP_B
 }
