+++ spring-client-kotlin-test/src/main/kotlin/com/markklim/libs/ginger/test/springclient/Application.kt
@@ -1,0 +1,12 @@
+package com.markklim.libs.ginger.test.springclient
+
+import com.markklim.libs.ginger.test.springclient.service.ApiService
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class Application
+
+    fun main(args: Array<String>) {
+        runApplication<Application>(*args).getBean(ApiService::class.java)
+    }
