+++ spring-client-kotlin-test/src/main/kotlin/com/markklim/libs/ginger/test/springclient/model/RequestModel.kt
@@ -1,0 +1,11 @@
+package com.markklim.libs.ginger.test.springclient.model
+
+data class RequestModel(
+    val param1: String = "param1",
+    val maskedParam: String = "param2",
+    val inner: RequestModelInnerClass = RequestModelInnerClass()
+)
+
+data class RequestModelInnerClass(
+    val param3: String = "param3",
+)
