--- spring-feign-kotlin-test/src/main/kotlin/com/markklim/libs/ginger/test/springfeign/configuration/ClientConfiguration.kt
@@ -1,27 +1,0 @@
-package com.markklim.libs.ginger.test.springfeign.configuration
-
-import com.markklim.libs.ginger.springfeign.WebClientLogging
-import com.markklim.libs.ginger.test.springfeign.client.ApiClient
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.reactive.function.client.support.WebClientAdapter
-import org.springframework.web.service.invoker.HttpServiceProxyFactory
-
-@Configuration
-class ClientConfiguration(@Value("\${client.api-client.url}") private val clientApiUrl: String) {
-
-    @Autowired(required = false)
-    private lateinit var webClientLogging: WebClientLogging
-
-    @Bean
-    fun apiClient(): ApiClient =
-        HttpServiceProxyFactory.builderFor(
-            WebClientAdapter.create(
-                webClientLogging.builder()
-                    .baseUrl(clientApiUrl)
-                    .build()
-            )
-        ).build().createClient(ApiClient::class.java)
-}
